refactory code and fix logs
